context_switch: Restore affinity and priority
The context_switch sets the CPU affinity to CPU 0 and the priority as
highest before starting the test, but doesn't restore those to the
original values.  On recent kernel, this could even result severe memory
pressure.  This commit fixes the problem by saving the original value
before starting the test and restoring after the test.

Signed-off-by: SeongJae Park <sjpark@amazon.de>